Move kube_proxy to tox/pytest
