+++ .travis.yml
@@ -127,8 +127,6 @@ env:
     - TRAVIS_FLAVOR=varnish FLAVOR_VERSION=4.1.7
     - TRAVIS_FLAVOR=varnish FLAVOR_VERSION=5.0.0
     - TRAVIS_FLAVOR=squid FLAVOR_VERSION=latest
-    - TRAVIS_FLAVOR=kube_proxy FLAVOR_VERSION=latest
-    - TRAVIS_FLAVOR=prometheus FLAVOR_VERSION=latest
     # END OF TRAVIS MATRIX
 install:
   - bundle install
