+++ circle.yml
@@ -83,8 +83,6 @@ test:
         - rake ci:run[kubernetes_state]
         - rake ci:run[cassandra_nodetool]
         - rake ci:run[squid]
-        - rake ci:run[kube_proxy]
-        - rake ci:run[prometheus]
         - bundle exec rake requirements
     post:
         - if [[ $(docker ps -a -q) ]]; then docker stop $(docker ps -a -q); fi
