+++ kube_proxy/MANIFEST.in
@@ -1,0 +1,5 @@
+include requirements.in
+include requirements.txt
+include tests/requirements.txt
+graft datadog_checks
+graft tests
