+++ kube_proxy/datadog_checks/kube_proxy/__init__.py
@@ -1,3 +1,6 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
 from . import kube_proxy
 
 KubeProxyCheck = kube_proxy.KubeProxyCheck
