--- kube_proxy/test/ci/kube_proxy.rake
@@ -1,65 +1,0 @@
-require 'ci/common'
-
-def kube_proxy_version
-  ENV['FLAVOR_VERSION'] || 'latest'
-end
-
-def kube_proxy_rootdir
-  "#{ENV['INTEGRATIONS_DIR']}/kube_proxy_#{kube_proxy_version}"
-end
-
-namespace :ci do
-  namespace :kube_proxy do |flavor|
-    task before_install: ['ci:common:before_install']
-
-    task :install do
-      Rake::Task['ci:common:install'].invoke('kube_proxy')
-      # sample docker usage
-      # sh %(docker create -p XXX:YYY --name kube_proxy source/kube_proxy:kube_proxy_version)
-      # sh %(docker start kube_proxy)
-    end
-
-    task before_script: ['ci:common:before_script']
-
-    task script: ['ci:common:script'] do
-      this_provides = [
-        'kube_proxy'
-      ]
-      Rake::Task['ci:common:run_tests'].invoke(this_provides)
-    end
-
-    task before_cache: ['ci:common:before_cache']
-
-    task cleanup: ['ci:common:cleanup']
-    # sample cleanup task
-    # task cleanup: ['ci:common:cleanup'] do
-    #   sh %(docker stop kube_proxy)
-    #   sh %(docker rm kube_proxy)
-    # end
-
-    task :execute do
-      exception = nil
-      begin
-        %w[before_install install before_script].each do |u|
-          Rake::Task["#{flavor.scope.path}:#{u}"].invoke
-        end
-        if !ENV['SKIP_TEST']
-          Rake::Task["#{flavor.scope.path}:script"].invoke
-        else
-          puts 'Skipping tests'.yellow
-        end
-        Rake::Task["#{flavor.scope.path}:before_cache"].invoke
-      rescue StandardError => e
-        exception = e
-        puts "Failed task: #{e.class} #{e.message}".red
-      end
-      if ENV['SKIP_CLEANUP']
-        puts 'Skipping cleanup, disposable environments are great'.yellow
-      else
-        puts 'Cleaning up'
-        Rake::Task["#{flavor.scope.path}:cleanup"].invoke
-      end
-      raise exception if exception
-    end
-  end
-end
