+++ kube_proxy/tests/fixtures/metrics_iptables.txt
