+++ kube_proxy/tests/fixtures/metrics_userspace.txt
